// src/services/types/api-endpoint-response.ts
/** biome-ignore-all lint/correctness/noUnusedImports: IDONCARE */
/** biome-ignore-all lint/suspicious/noExplicitAny: IDONCARE */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 * Source: api-docs
 * Title: CMS API
 * Description: CMS API
 * OpenAPI spec version: v1.0.0
 */

export const apiEndpointResponseHttpMethodsEnum = {GET: "GET", POST: "POST", PUT: "PUT", DELETE: "DELETE"} as const;

export type ApiEndpointResponseHttpMethodsEnumKey =
	(typeof apiEndpointResponseHttpMethodsEnum)[keyof typeof apiEndpointResponseHttpMethodsEnum];

export type ApiEndpointResponse = {
	/**
	 * @description API 엔드포인트 아이디
	 * @type integer, int64
	 */
	apiEndpointId: number;
	/**
	 * @description Tag name
	 * @type string
	 */
	tagName: string;
	/**
	 * @description Tag description
	 * @type string | undefined
	 */
	tagDescription?: string;
	/**
	 * @description 사용 가능한 HTTP 메서드 목록
	 * @type array
	 */
	httpMethods: ApiEndpointResponseHttpMethodsEnumKey[];
};